+++ nf_core/modules/patch.py
@@ -55,7 +55,6 @@ class ModulePatch(ModuleCommand):
         patch_relpath = Path(module_relpath, patch_filename)
         module_dir = Path(self.dir, module_relpath)
         patch_path = Path(self.dir, patch_relpath)
-        print(patch_path)
 
         if patch_path.exists():
             remove = questionary.confirm(
@@ -106,5 +105,4 @@ class ModulePatch(ModuleCommand):
 
         # Finally move the created patch file to its final location
         shutil.move(patch_temp_path, patch_path)
-        print(f"{patch_path} exists? {patch_path.exists()}")
         log.info(f"Patch file of '{module_fullname}' written to '{patch_path}'")